+++ examples/effects/pixellate.py
@@ -1,0 +1,24 @@
+"""Example of the punch effect."""
+import math
+from wasabi2d import run, Scene, event
+
+
+scene = Scene()
+
+logo = scene.layers[0].add_sprite(
+    'wasabi2d',
+    pos=(scene.width / 2, scene.height / 2),
+)
+effect = scene.layers[0].set_effect(
+    'pixellate',
+    pxsize=10
+)
+
+
+@event
+def update(dt, t):
+    effect.pxsize = 1 + round(20 * math.sin(t) ** 2)
+    logo.angle += 0.5 * dt
+
+
+run()