+++ docs/changelog.rst
@@ -7,6 +7,7 @@ History
 * New: add a :doc:`coroutine system <coros>`
 * New: make F12 :ref:`a built-in combination <screenshot>` for
   screenshots/video recording.
+* New: add a ``pixellate`` :doc:`post-processing effect <effects>`.
 
 
 1.2.0 - 2019-09-29